+++ Source/Injector/Pegasus.cs
@@ -15,7 +15,7 @@ using System.Windows.Forms;
 
 internal static class Program {
     internal const string DllName = "BedrockUtilityFramework.Xray.dll";
-    internal const string ExpectedHash = "E9ECF6F357F62224DD7BEC604F2A0F979C40934490572405462B27982FA6C487";
+    internal const string ExpectedHash = "52A2F98A40B0FD56971A2D06996DA0D5E0B0F23530C43D806CE5672C051AE7C5";
     internal static string DllPath { get { return Path.Combine(AppDomain.CurrentDomain.BaseDirectory, DllName); } }
     [STAThread] static int Main(string[] args) {
         if (args.Length == 1 && args[0] == "--test-host") { Thread.Sleep(30000); return 0; }